Move form-for tests to outside of src folder
