+++ packages/form-for/test/fixtures/Input.js
@@ -1,7 +1,7 @@
 // @flow
 
 import * as React from 'react';
-import type { ComponentProps } from '../../index';
+import type { ComponentProps } from '../../src';
 
 export default class Input extends React.Component<ComponentProps> {
   input: ?HTMLInputElement;
